Refactor Dockerfile based on developer guide
This dev container is non-systemd, we utilize a container entrypoint to configure and start up
services. The rails server must be foregrounded, I could not find a suitable rake task to start
it up in foreground, so at the moment we use rails server to bring it up but MIQ workers are not
getting started. This needs more information from MIQ team to be addressed or use supervisord.

Build summary :

- Install SCL in order to bring up PG 9.5 (PG 9.2 is not compatible with recent MIQ)
- Yum package list is the BARE MINIMUM to actually build and install MIQ with postgresql
- Defined ARGS and environment needed to bring up database in /etc/default/evm and /etc/default/evm_dev
- Absolutely no DB preparation is done during building, this happens in container-entrypoint
- We make two runs of bin/setup, one during building to install dependencies and a last one during startup
on container-entrypoint. The second run will actually prepare, migrate and seed the DB.

Is important to note that developer guide instructions heavily rely on running on a VM not a container.
Proper translations were made in order to follow the workflow as close as possible in a container.